% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tadaAICtable.R
\name{baselineSupport}
\alias{baselineSupport}
\title{Get the support for each baseline rate function from a tadaAICtable}
\usage{
baselineSupport(tadaAICtable)
}
\arguments{
\item{nbdaAICtable}{an object of class \code{\link{tadaAICtable}}.}
}
\value{
matrix giving the support (total Akaike Weight) and number of models fitted for each combination.
}
\description{
Calculates the support for each baseline rate function fitted in a set of models fitted using #'\code{\link{tadaAICtable}}.
}
\seealso{
\code{\link{oadaAICtable}}, \code{\link{tadaAICtable}}, \code{\link{typeSupport}},
\code{\link{networksSupport}}, \code{\link{modelAverageEstimates}}, \code{\link{variableSupport}},
\code{\link{unconditionalStdErr}}.
}
