% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tadaAICtable.R
\name{combineTadaAICtables}
\alias{combineTadaAICtables}
\title{Combine two or more tadaAICtables into a single tadaAICtable}
\usage{
combineTadaAICtables(tadaAICtableList, aicUse = "aicc",
  netComboModifier = rep("", length(tadaAICtableList)))
}
\arguments{
\item{tadaAICtableList}{a lift of \code{\link{tadaAICtable}} objects to be combined into a single table.}

\item{aicUse}{string specifying whether to use "aicc" or "aic".}

\item{netComboModifier}{optional character vector with length matching the length of \code{tadaAICtableList} to modify the
netcombo strings recorded from each individual \code{\link{tadaAICtable}} object (see below).}
}
\value{
An object of class \code{tadaAICtable}.
}
\description{
Takes two or more \code{\link{tadaAICtable}} objects, containing different models including the same number of networks and
the same ILVs, and combine them into a single table.
}
\details{
This function can be used for a variety of practical reasons. If the \code{\link{tadaAICtable}} is interrupted and a partial
tadaAICtable recovered, the remaining models can be run as a separate set and combined with the first partial set. Alternatively,
the user might want to add more models to a set without re-running the entire set, or run different parts of a model set on
different computers to speed up computation time. It is also possible to run model sets with different networks and then combine
them using \code{combineTadaAICtables}, so long as the number of networks matches (though it is possible to have dummy networks
that are constrained to have s=0 in all models, in order to match network number). In such cases a modifier can be added to the
netcombo codes from each tadaAICtable so they can be distinguished by functions such as \code{\link{networksSupport}}. e.g.
if we have two tadaAICtable objects each with the netcombos "1:0","0:1","1:2". If we use
\code{netComboModifier= c("NetworksA:","NetworksB:")} we get netcombos "NetworksA:1:0","NetworksA:0:1","NetworksA:1:2",
"NetworksB:1:0","NetworksB:0:1","NetworksB:1:2" in the resulting tadaAICtable object.
}
\seealso{
\code{\link{tadaAICtable}}
}
