% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filteredNBDAdata.R
\name{filteredNBDAdata}
\alias{filteredNBDAdata}
\title{Filter out individuals from an nbdaData object or dTADAData object.}
\usage{
filteredNBDAdata(nbdadata, filter, exclude)
}
\arguments{
\item{nbdadata}{an object of class \code{\link{nbdaData}} or \code{\link{dTADAData}}}

\item{filter}{a string specifying the part of the nbdaData or dTADAData object to use as a filter variable. Setting
\code{filter="id"} is the best easiest way to filter specific individuals (see \code{exclude} below). Alternatively, the user
could filter out individuals based on a specific ILV, e.g. \code{filter="asocILVdata[,1]"} uses the first asoc_ilv,
\code{filter="intILVdata[,2]"} uses the second int_ilv and  \code{filter="multiILVdata[,1]"} uses the first multi_ilv.}

\item{exclude}{a vector specifying the cases to be excluded. If "id" is used as a filter the target individuals can be specified
by concatenating as a string the label of the diffusion, "_", and the row number of the individual in assMatrix. e.g. if the
diffusion label is "Diffusion1" and we wish to filter out individuals 5 and 11, we would specify
\code{exclude=c("Diffusion1_5","Diffusion1_11")}. If we have specified a specific ILV as a filter, we provide a vector giving the
values of the ILV to be excluded. e.g. if we have a sex ILV specifying males=-0.5 and females=0.5, we can use
\code{exclude=-0.5} to exclude males from the analysis.}
}
\value{
An object of class \code{\link{nbdaData}} or \code{\link{dTADAData}} depending on the input data.
}
\description{
This function can be used to filter out individuals (or more generally lines of data). This removes them as learners only
so their influence as potential transmitters of information is not removed.
}
\details{
The user may wish to remove individuals from the NBDA to examine their influence on the results, or because they have good
reason to consider them to be unrepresentative outliers. Removing these individuals from the diffusion when creating the
\code{\link{nbdaData}} or \code{\link{dTADAData}} object has the disadvantage that it removes those individuals as potential
transmitters of information thus breaking up pathways of social transmission. Using \code{filteredNBDAdata} to remove
individuals retains their potential influence as transmitters of information but removes the as learners when fitting the model.
}
\seealso{
\code{\link{constrainedNBDAdata}}, \code{\link{nbdaData}}, \code{\link{dTADAData}}.
}
