% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oadaAICtable.R
\name{modelAverageEstimates}
\alias{modelAverageEstimates}
\title{Get the model averaged estimate for each variable from an oadaAICtable or tadaAICtable}
\usage{
modelAverageEstimates(nbdaAICtable, typeFilter = NULL,
  netFilter = NULL, baselineFilter = NULL, includeAsocial = TRUE,
  averageType = "mean")
}
\arguments{
\item{nbdaAICtable}{an object of class \code{\link{oadaAICtable}} or \code{\link{tadaAICtable}}.}

\item{typeFilter}{an optional string allowing the user to get the support for variables within a specified type of models.
e.g. \code{typeFilter="additive"} gets the variable support in the subset of additive models. See \code{\link{typeSupport}}
for an explanation of the different model types.}

\item{netFilter}{an optional string allowing the user to get the support for variables from models with a specific
combination of network effects, e.g. \code{netFilter = "0:1:0"} gets the variable support in the subset of models containing
only network 2. See \code{\link{networksSupport}} for an explanation of network combinations.}

\item{baselineFilter}{an optional string allowing the user to get the support for variables within the subset of models
with a specific baseline function, e.g. \code{typeFilter="gamma"} gets the variable support in the subset of models with a
gamma baseline rate function.
e.g. \code{typeFilter="additive"} gets the variable support in the subset of additive models.}

\item{includeAsocial}{logical indicating whether asocial models should be included.}

\item{averageType}{string indicating whether a "mean" or "median" should be calculated.}
}
\value{
numeric vector giving the model averaged estimate for each variable.
}
\description{
Calculates the model averaged estimate for each variable fitted in a set of models fitted using \code{\link{oadaAICtable}}
or code{\link{tadaAICtable}}.
}
\details{
By default the model averaged estimate is calculated as the weighted mean of maximum likelihood estimates in each model,
weighted by Akaike weights. This can be changed to a weighted median using the \code{averageType} argument. This could be
useful in cases where a few models (potentially of low weight) have an essentially infinite estimate for a model parameter,
e.g. this can occur for s parameters in an OADA when the diffusion follows the network closely. In such cases a Akaike
weighted median may better reflect the overall findings of the multi-model analysis. Models where a variable does not appear
are treated as models in which the variable has a MLE=0.
}
\seealso{
\code{\link{oadaAICtable}}, \code{\link{tadaAICtable}}, \code{\link{typeSupport}},
\code{\link{typeByNetworksSupport}}, \code{\link{networksSupport}}, \code{\link{modelAverageEstimates}},
\code{\link{unconditionalStdErr}}.
}
