% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oadaAICtable.R
\name{multiModelLowerLimits_multicore}
\alias{multiModelLowerLimits_multicore}
\title{Test the sensitivity of the lower limit of an s parameter to model selection uncertainty}
\usage{
multiModelLowerLimits_multicore(which, aicTable, cores = 2,
  deltaThreshold = Inf, conf = 0.95, modelIndex = NULL,
  searchRange = NULL, exclude.innovations = T, innovations = NULL,
  startValue = NULL, lowerList = NULL, upperList = NULL,
  method = "nlminb", gradient = T, iterations = 150)
}
\arguments{
\item{which}{numeric giving the parameter for which the confidence interval is to be calculated. The appropriate number
can be identified from the fitted model, by entering \code{<modelName>@varNames} to extract the variable names from
the model. Each variable name is preceded by its number.}

\item{aicTable}{an object of class \code{\link{oadaAICtable}} or \code{\link{tadaAICtable}}.}

\item{cores}{numerical giving the number of computer cores to be used in parallel to fit the models in the set, thus speeding
up the process. By default set to 2. For a standard desktop computer at the time of writing 4-6 is advised.}

\item{deltaThreshold}{optional numerical determining the threshold difference in AICc/AIC for a model to be included in the
output. e.g. \code{deltaThreshold=10} includes all models within 10 AICc units of the best model.}

\item{conf}{numerical giving the level of confidence required, defaulting to the traditional 0.95.}

\item{modelIndex}{optional numeric vector specifiying which models to include in the output, subject to \code{deltaThreshold}.}

\item{searchRange}{optional numeric vector of length two, giving the range within which to search for the lower endpoint. If
omitted, the function searches between 0 and the MLE for s in each model.}

\item{exclude.innovations}{logical determining whether innovation events (the first individual to learn in each diffusion)
should be excluded from the calculation- since we know the innovation events must occur by asocial learning not social
transmission.}

\item{innovations}{numerical giving the number of innovations across all diffusions. By default this is assumed to be one
innovator per diffusion in which there were no demonstrators (see \code{demons} argument in \code{\link{nbdaData}})}

\item{startValue}{optional numeric vector giving start values for the maximum likelihood optimization. Length to match
the number of parameters fitted in the full model.}

\item{lowerList}{optional numeric matrix giving lower values for the maximum likelihood optimization for each model.
Columns to match the number of parameters fitted in the full model, rows matched to the number of models. Can be used if
some models have convergence problems or trigger errors.}

\item{upperList}{optional numeric matrix giving upper values for the maximum likelihood optimization for each model.
Columns to match the number of parameters fitted in the full model, rows matched to the number of models. Can be used if
some models have convergence problems or trigger errors.}

\item{method}{optional character string passed to \code{\link{oadaFit}}.}

\item{gradient}{optional logical passed to \code{\link{oadaFit}}.}

\item{iterations}{optional numerical passed to \code{\link{oadaFit}}.}
}
\value{
data.frame. See \code{\link{multiModelLowerLimits}} for details.
}
\description{
Runs \code{\link{multiModelLowerLimits}} across multiple computer cores. see documentation for
\code{\link{multiModelLowerLimits}} for details.
}
\seealso{
\code{\link{multiModelLowerLimits}}
}
