% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oadaAICtable.R
\name{multiModelPropST}
\alias{multiModelPropST}
\title{Obtain estimates of the proportion of events that occured by social transmission for a set of models}
\usage{
multiModelPropST(aicTable, subset = NULL, statusBar = T)
}
\arguments{
\item{aicTable}{an object of class \code{\link{oadaAICtable}} or \code{\link{tadaAICtable}}.}

\item{subset}{optional numerical vector specifying a subset of model numbers to be included.}

\item{statusBar}{logical indictaing whether a status/ progress bar should be produced}
}
\value{
list with the following components:\describe{
  \item{$propSTtable}{propST via each network for each model in the set}
  \item{$modelAverages}{model averaged propST for each network}
  }
}
\description{
Takes a set of models fitted by \code{\link{oadaAICtable}} or \code{\link{tadaAICtable}} and calculates the estimated
proportion of events occuring by social transmssion via each network (propST) using \code{\link{nbdaPropSolveByST}} in each
model. This, alongside the \code{\link{multiModelLowerLimits}} function, allows the user to test the robustness of conclusions
about social transmission to model selection uncertainty. The function also calculates a model averaged estimate of propST for
each network.
}
\seealso{
\code{\link{multiModelLowerLimits}}, \code{\link{oadaAICtable}}, \code{\link{tadaAICtable}},
\code{\link{plotProfLik}}, \code{\link{nbdaPropSolveByST}}, \code{\link{multiModelLowerLimits}}.
}
