% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oadaPropSolveByST.R
\name{nbdaPropSolveByST.byevent}
\alias{nbdaPropSolveByST.byevent}
\alias{oadaPropSolveByST.byevent}
\title{Estimate the probability that each event occured by social transmission}
\usage{
nbdaPropSolveByST.byevent(par = NULL, nbdadata = NULL, model = NULL,
  type = "social", retainInt = NULL)
}
\arguments{
\item{par}{optional numeric giving the parameter values. Not necessary if \code{model} is specified.}

\item{nbdadata}{object of class \code{\link{nbdaData}}, list of objects of class \code{\link{nbdaData}}, object of class
\code{\link{dTADAData}} or list of objects of class \code{\link{dTADAData}}. Not necessary if \code{model} is specified.}

\item{model}{object of class \code{\link{oadaFit}} or \code{\link{tadaFit}}.}

\item{type}{string determining whether the estimates are for an "asocial" or "social" model. This is taken from
\code{model} if provided.}

\item{retainInt}{logical, can be used to force the model to retain int_ilvs in an asocial model. This is used internally
by other functions when there is an offset on the s parameters, but can usually be safely ignored by the user.}
}
\value{
A dataframe giving the predicted probability that each event occured by social transmission via each network.
}
\description{
Generates a table giving the predicted probability that each event occured by social transmission via each network.
}
\details{
The user can provide the name of a model- in which case the MLEs of that model are input as parameter values. Instead the
user can input their own parameter values in which case \code{nbdadata} must also be provided. The function works for
multiple diffusions. A column is also provided for P(S offset), which gives the probability that transmission was via
network effects that have been allocated to the offset fo s parameters, see \code{\link{constrainedNBDAdata}}.
}
\seealso{
\code{\link{nbdaPropSolveByST}}
}
