% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oadaAICtable.R
\name{networksSupport}
\alias{networksSupport}
\title{Get the support for each network combination from an oadaAICtable or tadaAICtable}
\usage{
networksSupport(nbdaAICtable)
}
\arguments{
\item{nbdaAICtable}{an object of class \code{\link{oadaAICtable}} or \code{\link{tadaAICtable}}.}
}
\value{
dataframe giving the support (total Akaike Weight) and number of models fitted for each combination.
}
\description{
Calculates the support for network combination fitted in a set of models fitted using \code{\link{oadaAICtable}} or
\code{\link{tadaAICtable}}.
}
\details{
Network combinations are represented by the set of constraints on the s parameters used to generate the model, using the
\code{constraintsVectMatrix} argument. e.g. 1:1:1 means all three networks in the data were constrained to have the same
effect, 1:2:0 means the first two networks have different effects, whereas the third is constrained to have no effect, and
0:0:0 represents asocial models.
}
\seealso{
\code{\link{oadaAICtable}}, \code{\link{tadaAICtable}}, \code{\link{typeSupport}},
\code{\link{networksSupport}}, \code{\link{modelAverageEstimates}}, \code{\link{variableSupport}},
\code{\link{unconditionalStdErr}}.
}
