% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oadaAICtable.R
\name{oadaAICtable}
\alias{oadaAICtable}
\title{Fit a set of OADA models for multi-model inference}
\usage{
oadaAICtable(nbdadata, constraintsVectMatrix, typeVect = NULL,
  offsetVectMatrix = NULL, cores = 1, modelsPerCorePerSet = NULL,
  writeProgressFile = F, saveTableList = F, statusBar = NULL,
  startValue = NULL, method = "nlminb", gradient = T,
  iterations = 150, aicUse = "aicc", lowerList = NULL,
  upperList = NULL, combineTables = F, MLEs = NULL, SEs = NULL,
  MLEilv = NULL, SEilv = NULL, MLEint = NULL, SEint = NULL,
  convergence = NULL, loglik = NULL, aic = NULL, aicc = NULL,
  netComboModifierVect = "", saveModels = F, stripData = F,
  models = NULL)
}
\arguments{
\item{nbdadata}{an object of class (\code{\link{nbdaData}}) to fit models to a single diffusion or a list of
nbdaData objects to fit a model to multiple diffusions.}

\item{constraintsVectMatrix}{a numerical matrix specifying the constraints, with each row specifiying a model to be fitted.
The number of columns is equal to the number of parameters in the (\code{\link{nbdaData}}) object(s) input to argument
\code{nbdadata}. Each row then specifies the \code{constraintsVect} to be passed to the \code{\link{constrainedNBDAdata}}
when creating the data object(s) to which that model is fitted.}

\item{typeVect}{optional character vector specifying if each model is "asocial" or "social". However, it is not usually
necessary to specify, since models with all s parameters constrained =0 are automatically classified as "asocial", and
others are assumed to be "social".}

\item{offsetVectMatrix}{an optional numerical matrix specifying the offsets, with each row specifiying the offsets for each
model to be fitted. The number of columns is equal to the number of parameters in the (\code{\link{nbdaData}}) object(s)
input to argument \code{nbdadata}. Each row then specifies the \code{offsetVect} to be passed to the
\code{\link{constrainedNBDAdata}} when creating the data object(s) to which that model is fitted.}

\item{cores}{numerical giving the number of computer cores to be used in parallel to fit the models in the set, thus
speeding up the process. By default set to 1. For a standard desktop computer at the time of writing 4-6 is advised.}

\item{modelsPerCorePerSet}{optional numerical. If specified the models can be fit in sets, and a progress file written
after each set is completed. This means progress is not completely lost in the case of a crash/ powercut etc. For example,
if we have 400 models, we can specify \code{cores=4} and \code{modelsPerCorePerSet=10}. This means 10 models are fitted on
each core, then progress is saved with the first 40 models, then the next 40 and so on.}

\item{writeProgressFile}{logical. If set to T, a file is written to the working directory when each set of models have
been completed with the oadaAICtable for the models fitted so far. In the event of a crash, the remining models can be
fitted as a separate set, then combined using \code{\link{combineOadaAICtables}}.}

\item{saveTableList}{logical. If set to T, a file is written to the working directory containing a list of oadaAICtable
objects- one for each set of models per core (see \code{modelsPerCorePerSet}). This is useful if there is an error in one
or more of the models preventing the oadaAICtable from being calculated properly. In this case the set(s) containing the
faulty model can be corrected, refitted using \code{oadaAICtable}, and replaced in the list. The oadaAICtable can then
be reformed using \code{\link{combineOadaAICtables}}.}

\item{statusBar}{optional logical. Status bar only works when \code{cores=1}.}

\item{startValue}{optional numeric vector giving start values for the maximum likelihood optimization. Length to match
the number of parameters fitted in the full model.}

\item{method}{optional character string passed to \code{\link{oadaFit}}.}

\item{gradient}{optional logical passed to \code{\link{oadaFit}}.}

\item{iterations}{optional numerical passed to \code{\link{oadaFit}}.}

\item{aicUse}{string specifying whether to use "aicc" or "aic".}

\item{lowerList}{optional numeric matrix giving lower values for the maximum likelihood optimization for each model.
Columns to match the number of parameters fitted in the full model, rows matched to the number of models. Can be used if
some models have convergence problems or trigger errors.}

\item{upperList}{optional numeric matrix giving upper values for the maximum likelihood optimization for each model.
Columns to match the number of parameters fitted in the full model, rows matched to the number of models. Can be used if
some models have convergence problems or trigger errors.}

\item{combineTables}{logical used internally by \code{\link{combineOadaAICtables}} when making calls to \code{oadaAICtable}.}

\item{saveModels}{logical determining whether the fitted model objects should be saved as a list as a component in the
\code{oadaAICtable} object. This can be useful for accessing individual models without re-fitting them (e.g. to get
confidence intervals using \code{\link{profLikCI}}). Alternatively individual problem models can be refitted, replaced in
the list, and the \code{oadaAICtable} re-generated using \code{oadaAICtable} using the \code{models} argument (see below).}

\item{stripData}{logical. By default, when \code{saveModels=T} each model is saved with its constrained version of the
data. Whilst this can be useful, if there are lots of models and/or a lot of data, this can take up too much memory.
setting \code{saveModels=T} and \code{stripData=T} enables the models to be saved without their data.}

\item{models}{an optional list of model objects of class \code{\link{oadaFit}}. If provided, the models are not fitted
but taken from the list when generating the oadaAICtable. This is useful if individual models have been corrected and the
user needs to re-form the oadaAICtable.}
}
\value{
An object of class \code{oadaAICtable}.
}
\description{
\code{oadaAICtable} takes diffusion data in the form of an nbdaData object (\code{\link{nbdaData}}) or a list of nbdaData
objects (for multiple diffusions). It then fits a set of models using \code{\link{oadaFit}} and return them in an object
of class \code{oadaAICtable}. Arguments not described below are used internally by \code{\link{combineOadaAICtables}} when
making calls to \code{oadaAICtable} and can be ignored by the user.
}
\details{
Each row of \code{constraintsVectMatrix} and \code{offsetVectMatrix} determines a model to
be fitted. For each row, constrained \code{nbdaData} objects are created using (\code{\link{constrainedNBDAdata}}) with
\code{constraintsVect=constraintsVectMatrix[i,]} and \code{offsetVect=offsetVectMatrix[i,]}. A model is then fitted to
the \code{nbdaData} object(s) using \code{\link{oadaFit}}.
}
\section{print(oadaAICtable)  components}{

A data.frame giving a summary of models ordered by AIC can be obtained using \code{print(<name of oadaAICtable>)}. This has
the following columns, listed in order: \describe{
  \item{model}{Model number, i.e. the row of \code{constraintsVectMatrix} used to generate the model.}
  \item{type}{Type of model. noILVs, additive (ILV effects on asocial learning only), multiplicative (all ILVs have same
  effect on asocial and social learning), unconstrained (differing effects on asocial and social learning for at least one
  ILV), socialEffectsOnly (ILV effects only on social learning),asocial}
  \item{netcombo}{A representaion of the network effects present in the model, i.e. the constraints on the s
  parameters. See \code{\link{constrainedNBDAdata}}.}
  \item{baseline}{Used for TADA models only, set to NA here.}
  \item{CONS.}{The constraint on each parameter, as taken from \code{constraintsVectMatrix}}
  \item{OFF}{The offset on each parameter, as taken from \code{offsetVectMatrix}}
  \item{convergence}{Was convergence reported by the optimization algorithm?}
  \item{loglik}{-log-likelihood for the model.}
  \item{s...}{Maximum likelihood estimates for s parameters.}
  \item{ASOCIAL...}{Maximum likelihood estimates for effects of ILVs on asocial learning.}
  \item{SOCIAL...}{Maximum likelihood estimates for effects of ILVs on social learning.}
  \item{ASOCIAL:SOCIAL...}{Maximum likelihood estimates for multiplicative effects of ILVs see \code{\link{oadaFit}}.}
  \item{SE...}{Standard errors for each parameter, set to 0 when a parameter was constrained. See \code{\link{oadaFit}}.}
  \item{aic}{AIC for the model.}
  \item{aicc}{AICc for the model.}
  \item{deltaAICc}{Difference in AICc or AIC from the best model.}
  \item{RelSupport}{Relative support for the model compared to the best model, calculated as exp(-0.5*deltaAICc).}
  \item{AkaikeWeight}{Akaike weight for the model. Can be interpretted as the probability that model has the highest
  predictive power (K-L information) out of the set of models considered.}
  }
}

\section{oadaAICtable components}{
\describe{
  \item{@nbdadata}{The unconstrained data the model is fitted to, as a list of nbdaData objects.}
  \item{@models}{A list of fitted model objects of class \code{\link{oadaFit}}, if saved.}
  \item{@convergence}{Was convergence reported by the optimization algorithm? (Ordered by \code{constraintsVectMatrix}).}
  \item{@logLik}{-log-likeihood for models (ordered by \code{constraintsVectMatrix}).}
  \item{@aicc}{AICc for models (ordered by \code{constraintsVectMatrix}).}
  \item{@aic}{AIC for models (ordered by \code{constraintsVectMatrix}).}
  \item{@constraintsVectMatrix}{\code{constraintsVectMatrix} input to the function.}
  \item{@offsetVectMatrix}{\code{offsetVectMatrix} input to the function.}
  \item{@MLEs}{Maximum likelihood estimates for s parameters (ordered by \code{constraintsVectMatrix}).}
  \item{@SEs}{Standard errors for s parameters (ordered by \code{constraintsVectMatrix}).}
  \item{@MLEilv}{Maximum likelihood estimates for effect of ILs on asocial learning (ordered by
  \code{constraintsVectMatrix}).}
  \item{@SEilv}{Standard errors for effect of ILs on asocial learning  (ordered by \code{constraintsVectMatrix}).}
  \item{@MLEint}{Maximum likelihood estimates for effect of ILs on social learning (ordered by
  \code{constraintsVectMatrix}).}
  \item{@SEint}{Standard errors for effect of ILs on social learning  (ordered by \code{constraintsVectMatrix}).}
  \item{@typeVect}{Type of models (ordered by \code{constraintsVectMatrix}).}
  \item{@deltaAICc}{Difference in AICc or AIC from the best model. (ordered by \code{constraintsVectMatrix}).}
  \item{@RelSupport}{Relative support for the model compared to the best model, calculated as exp(-0.5*deltaAICc).
  (ordered by \code{constraintsVectMatrix}).}
  \item{@RelSupport}{Akaike weight for the model. (ordered by \code{constraintsVectMatrix}).}
  \item{@printTable}{data.frame to be output by the print method for \code{oadaAICtable} (see above).}
}
}

\seealso{
\code{\link{typeSupport}}, \code{\link{networksSupport}}, \code{\link{typeByNetworksSupport}}, \code{\link{modelAverageEstimates}},
\code{\link{variableSupport}}, \code{\link{unconditionalStdErr}}, \code{\link{combineOadaAICtables}}. For TADA models
use \code{\link{tadaAICtable}}.
}
