% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotConnections.R
\name{plotConnections}
\alias{plotConnections}
\title{Plot the connections to informed individuals against event order}
\usage{
plotConnections(nbdadata, network = 1, rescale = F, lty = 1,
  lwd = 1, col = 2, pchVector = NULL, symbol = NULL,
  xlab = "Acquisition event", ylab = NULL, title = NULL,
  plotID = NULL, offset = c(0.1, 0), xlim = NULL, ylim = NULL,
  titlePos = c(0, 0), averageLinelty = 2, averageLinecol = 4,
  averageLinelwd = 1, plotConnectLine = T, plotAverageLine = T)
}
\arguments{
\item{nbdadata}{an object of class \code{\link{nbdaData}}.}

\item{network}{numeric giving the number of the network for which the connections are to be plotted, i.e. corresponding to
the third dimension of assMatrix in the \code{\link{nbdaData}} object.}

\item{rescale}{logical indicating whether the connections should be rescaled to be plotted relative to the average
connection to informed individuals for each event.}

\item{lty}{numeric or string determining the connecting line type, see (\code{\link[graphics]{par}}).}

\item{lwd}{numeric determining the connecting line width, see (\code{\link[graphics]{par}}).}

\item{pchVector}{an optional vector of plotting symbols to be used for each type of individual as defined by \code{symbol}
below. See \code{pch} in (\code{\link[graphics]{par}}).}

\item{symbol}{an optional vector allowing different symbols to be plotted for each point. e.g. to plot different symbols for
learners an non learners, specify \code{symbol=<name of nbdadata>@status}. To choose which symbol is used for each category,
use \code{symbol}, e.g. \code{pchVector = c(3,8)}. Alternatively, the user could plot points based on a specific ILV, e.g.
\code{symbol=<name of nbdadata>@asocILVdata[,1]} to the first asoc_ilv.}

\item{xlab}{passed to the \code{\link[graphics]{plot}} function.}

\item{ylab}{passed to the \code{\link[graphics]{plot}} function.}

\item{title}{an optional string giving a title for the plot.}

\item{plotID}{an optional vector of names or indentifiers to be plotted next to the symbol for each learner. If \code{idname} was
provided to the \code{\link{nbdaData}} function, then this can be done by specifiying
\code{plotID = <name of nbdadata>@idname[<name of nbdadata>@orderAcq]}}

\item{offset}{an optional numerical vector of length two, specifying where to plot id relative to its corresponding point.}

\item{xlim}{passed to the \code{\link[graphics]{plot}} function.}

\item{ylim}{passed to the \code{\link[graphics]{plot}} function.}

\item{titlePos}{an optional numerical vector of length two, specifying where to plot the title.}

\item{averageLinelty}{numeric or string determining the average line type, see (\code{\link[graphics]{par}}).}

\item{averageLinecol}{numeric or string determining the average line colour, see (\code{\link[graphics]{par}}).}

\item{averageLinelwd}{numeric or string determining the average line colour, see (\code{\link[graphics]{par}}).}

\item{plotConnectLine}{logical specifying whether a connecting line is to be plotted.}

\item{plotAverageLine}{logical specifying whether an average line is to be plotted.}
}
\description{
For each acquisition event, the total connection of each naive individual to informed individuals is plotted for a specified
network.
}
\details{
The individual to learn each event is plotted in red and these are connected by a red line (by default). A dotted
line showing the average connection to informed individuals each event is plotted in blue (by default). If the order of
acquisition does not follow the network in question, we would expect the red points to be distributed evenly above and
below the blue line. The more closely the order of acquisition follows the network, the more points will be above the blue
line and the higher they will be above it.
}
\seealso{
\code{\link{nbdaData}}
}
