% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profLikCI.R
\name{plotProfLikTrueTies}
\alias{plotProfLikTrueTies}
\title{Plot profile likelihood for a parameter in an OADA with trueTies}
\usage{
plotProfLikTrueTies(which, model, range, resolution = 20,
  inflation = 1, conf = 0.95, startValue = NULL, lower = NULL,
  upper = NULL)
}
\arguments{
\item{which}{numeric giving the parameter for which the profile likelihood is to be plotted. The appropriate number
can be identified from the fitted model, by entering \code{<modelName>@varNames} to extract the variable names from
the model. Each variable name is preceded by its number.}

\item{model}{object of class \code{\link{oadaFit}}.}

\item{range}{numeric vector of length two, providing a range of parameter values within which the profile likelihood is
to be plotted.}

\item{resolution}{numeric giving the number of points to be plotted. The user is advised to start at a low resolution to
obtain the approproate range, then increase resolution to identify ranges for endpoints.}

\item{inflation}{numerical to be used if the confidence intervals are to be inflated by a specified amount, as suggested
by Burnham & Anderson (2000) to allow for model selection uncertainty. This simply increases the height of the dotted
line above the model maximum likelihood by a factor of \code{inflation}.}

\item{conf}{numerical giving the level of confidence required, defaulting to the traditional 0.95.}

\item{lower}{optional numeric vector giving lower values for the maximum likelihood optimization. Length to match
the number of parameters fitted in the model. By default taken to be 0 for all s parameters and -Inf for coefficients of
ILVs.}

\item{upper}{optional numeric vector giving upper values for the maximum likelihood optimization. Length to match
the number of parameters fitted in the model. By default taken to be Inf for all parameters.}
}
\value{
A dataframe giving the plotted values.
}
\description{
Find the profile likelihood for a specific parameter in an OADA with trueTies
}
\details{
An alternative simple plotting function to plot profile likelihoods, intended for data with trueTies, only works for social
models. Functions in much the same way as \code{\link{plotProfLik}}, except it cannot be used to find the profile likelihood
for the difference between two parameters. For this use \code{\link{plotProfLikDiffTrueTies}}.
}
\seealso{
\code{\link{plotProfLik}}
}
