% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profLikCI.R
\name{profLikCIDiffTrueTies}
\alias{profLikCIDiffTrueTies}
\title{Find confidence intervals for the difference between two parameters in an OADA with trueTies}
\usage{
profLikCIDiffTrueTies(which, whichBaseline, model, interval,
  inflation = 1, conf = 0.95, startValue = NULL, lower = NULL,
  upper = NULL)
}
\arguments{
\item{which}{numeric giving the first parameter for the difference. The appropriate number
can be identified from the fitted model, by entering \code{<modelName>@varNames} to extract the variable names from
the model. Each variable name is preceded by its number.}

\item{whichBaseline}{numeric giving the second parameter for the difference. The confidence interval is for which -
whichBaseline.}

\item{model}{object of class \code{\link{oadaFit}}.}

\item{interval}{numeric vector of length two, providing a range within which the required endpoint of the confidence interval
is known to be located. This range can be indentified using \code{\link{plotProfLikTrueTies}}.}

\item{inflation}{numerical to be used if the confidence intervals are to be inflated by a specified amount, as suggested
by Burnham & Anderson (2000) to allow for model selection uncertainty. This simply increases the height of the dotted
line above the model maximum likelihood by a factor of \code{inflation}.}

\item{conf}{numerical giving the level of confidence required, defaulting to the traditional 0.95.}

\item{startValue}{optional numeric vector giving start values for the maximum likelihood optimization. Length to match
the number of parameters fitted in the model.}

\item{lower}{optional numeric vector giving lower values for the maximum likelihood optimization. Length to match
the number of parameters fitted in the model. By default taken to be 0 for all s parameters and -Inf for coefficients of
ILVs.}

\item{upper}{optional numeric vector giving upper values for the maximum likelihood optimization. Length to match
the number of parameters fitted in the model. By default taken to be Inf for all parameters.}
}
\value{
numeric giving the location of the required endpoint.
}
\description{
Find the confidence intervals for the difference between two parameters in an OADA with trueTies
}
\details{
An alternative to find confidence intervals, intended for data with trueTies, only works for socialmodels. Functions in much
the same way as \code{\link{profLikCI}}, except the upper and lower endpoints need to be found separately.
}
\seealso{
\code{\link{profLikCI}},\code{\link{plotProfLikTrueTies}}, \code{\link{plotProfLikDiffTrueTies}}
}
