% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tadaAICtable.R
\name{tadaAICtable}
\alias{tadaAICtable}
\title{Fit a set of TADA models for multi-model inference}
\usage{
tadaAICtable(nbdadata, constraintsVectMatrix, typeVect = NULL,
  baselineVect = NULL, offsetVectMatrix = NULL, cores = 1,
  modelsPerCorePerSet = NULL, writeProgressFile = F,
  statusBar = NULL, noHazFunctParsCustom = NULL,
  hazFunct = function() return(NULL), cumHaz = function() return(NULL),
  startValue = NULL, method = "nlminb", gradient = T,
  iterations = 150, aicUse = "aicc", lowerList = NULL,
  combineTables = F, MLEs = NULL, SEs = NULL, MLEilv = NULL,
  SEilv = NULL, MLEint = NULL, SEint = NULL, MLEhaz = NULL,
  SEhaz = NULL, convergence = NULL, loglik = NULL, aic = NULL,
  aicc = NULL, netComboModifierVect = "")
}
\arguments{
\item{nbdadata}{an object of class (\code{\link{nbdaData}}) or (\code{\link{dTADAData}}) to fit models to a single diffusion or a list of
nbdaData/dTADAData objects to fit a model to multiple diffusions.}

\item{constraintsVectMatrix}{a numerical matrix specifying the constraints, with each row specifiying a model to be fitted.
The number of columns is equal to the number of parameters in the (\code{\link{nbdaData}}) or (\code{\link{dTADAData}}) object(s) input to argument
\code{nbdadata}. Each row then specifies the \code{constraintsVect} to be passed to the \code{\link{constrainedNBDAdata}}
when creating the data object(s) to which that model is fitted.}

\item{typeVect}{optional character vector specifying if each model is "asocial" or "social". However, it is not usually
necessary to specify, since models with all s parameters constrained =0 are automatically classified as "asocial", and
others are assumed to be "social".}

\item{baselineVect}{optional character vector specifying the baseline function for each model (see \code{\link{tadaFit}}).}

\item{offsetVectMatrix}{an optional numerical matrix specifying the offsets, with each row specifiying the offsets for each
model to be fitted. The number of columns is equal to the number of parameters in the (\code{\link{nbdaData}}) or (\code{\link{dTADAData}}) object(s)
input to argument \code{nbdadata}. Each row then specifies the \code{offsetVect} to be passed to the
\code{\link{constrainedNBDAdata}} when creating the data object(s) to which that model is fitted.}

\item{cores}{numerical giving the number of computer cores to be used in parallel to fit the models in the set, thus
speeding up the process. By default set to 1. For a standard desktop computer at the time of writing 4-6 is advised.}

\item{modelsPerCorePerSet}{optional numerical. If specified the models can be fit in sets, and a progress file written
after each set is completed. This means progress is not completely lost in the case of a crash/ powercut etc. For example,
if we have 400 models, we can specify \code{cores=4} and \code{modelsPerCorePerSet=10}. This means 10 models are fitted on
each core, then progress is saved with the first 40 models, then the next 40 and so on.}

\item{writeProgressFile}{logical. If set to T, a file is written to the working directory when each set of models have
been completed with the tadaAICtable for the models fitted so far. In the event of a crash, the remining models can be
fitted as a separate set, then combined using \code{\link{combineTadaAICtables}}.}

\item{statusBar}{optional logical. Status bar only works when \code{cores=1}.}

\item{noHazFunctParsCustom}{optional numerical necessary if "custom" is specified for any models in \code{baselineVect}.
See \code{\link{tadaFit}} for details.}

\item{hazFunct}{optional function necessary if "custom" is specified for any models in \code{baselineVect}.
See \code{\link{tadaFit}} for details.}

\item{cumHaz}{optional function necessary if "custom" is specified for any models in \code{baselineVect}.
See \code{\link{tadaFit}} for details.}

\item{startValue}{optional numeric vector giving start values for the maximum likelihood optimization. Length to match
the number of parameters fitted in the full model.}

\item{method}{optional character string passed to \code{\link{tadaFit}}.}

\item{gradient}{optional logical passed to \code{\link{tadaFit}}.}

\item{iterations}{optional numerical passed to \code{\link{tadaFit}}.}

\item{aicUse}{string specifying whether to use "aicc" or "aic".}

\item{lowerList}{optional numeric matrix giving lower values for the maximum likelihood optimization for each model.
Columns to match the number of parameters fitted in the full model, rows matched to the number of models. Can be used if
some models have convergence problems or trigger errors.}

\item{combineTables}{logical used internally by \code{\link{combineTadaAICtables}} when making calls to \code{tadaAICtable}.}

\item{upperList}{optional numeric matrix giving upper values for the maximum likelihood optimization for each model.
Columns to match the number of parameters fitted in the full model, rows matched to the number of models. Can be used if
some models have convergence problems or trigger errors.}
}
\value{
An object of class \code{tadaAICtable}.
}
\description{
\code{tadaAICtable} takes diffusion data in the form of an nbdaData object (\code{\link{nbdaData}}) or (\code{\link{dTADAData}})
or a list of nbdaData/ dTADAData objects (for multiple diffusions). It then fits a set of models using \code{\link{tadaFit}}
and return them in an object of class \code{tadaAICtable}. Arguments not listed below are used internally by
\code{\link{combineTadaAICtables}} when making calls to \code{tadaAICtable} and can be ignored by the user.
}
\details{
Each row of \code{constraintsVectMatrix}, \code{offsetVectMatrix}, and \code{baselineVect} determines a model to
be fitted. For each row, constrained \code{nbdaData} objects are created using (\code{\link{constrainedNBDAdata}}) with
\code{constraintsVect=constraintsVectMatrix[i,]} and \code{offsetVect=offsetVectMatrix[i,]}. A model is then fitted to
the \code{nbdaData} object(s) using \code{\link{tadaFit}} with a baseline determined by \code{offsetVect=baselineVect[i]}.
}
\section{print(tadaAICtable)  components}{

A data.frame giving a summary of models ordered by AIC can be obtained using \code{print(<name of tadaAICtable>)}. This has
the following columns, listed in order: \describe{
  \item{model}{Model number, i.e. the row of \code{constraintsVectMatrix} used to generate the model.}
  \item{type}{Type of model. noILVs, additive (ILV effects on asocial learning only), multiplicative (all ILVs have same
  effect on asocial and social learning), unconstrained (differing effects on asocial and social learning for at least one
  ILV), asocial}
  \item{netcombo}{A representaion of the network effects present in the model, i.e. the constraints on the s
  parameters. See \code{\link{constrainedNBDAdata}}.}
  \item{baseline}{The baseline function used for each model.}
  \item{CONS.}{The constraint on each parameter, as taken from \code{constraintsVectMatrix}}
  \item{OFF}{The offset on each parameter, as taken from \code{offsetVectMatrix}}
  \item{convergence}{Was convergence reported by the optimization algorithm?}
  \item{loglik}{-log-likelihood for the model.}
  \item{s...}{Maximum likelihood estimates for s parameters.}
  \item{ASOCIAL...}{Maximum likelihood estimates for effects of ILVs on asocial learning.}
  \item{SOCIAL...}{Maximum likelihood estimates for effects of ILVs on social learning.}
  \item{ASOCIAL:SOCIAL...}{Maximum likelihood estimates for multiplicative effects of ILVs see \code{\link{tadaFit}}.}
  \item{SE...}{Standard errors for each parameter, set to 0 when a parameter was constrained. See \code{\link{tadaFit}}.}
  \item{aic}{AIC for the model.}
  \item{aicc}{AICc for the model.}
  \item{deltaAICc}{Difference in AICc or AIC from the best model.}
  \item{RelSupport}{Relative support for the model compared to the best model, calculated as exp(-0.5*deltaAICc).}
  \item{AkaikeWeight}{Akaike weight for the model. Can be interpretted as the probability that model has the highest
  predictive power (K-L information) out of the set of models considered.}
  }
}

\section{tadaAICtable components}{
\describe{
  \item{@nbdadata}{The unconstrained data the model is fitted to, as a list of nbdaData objects.}
  \item{@convergence}{Was convergence reported by the optimization algorithm? (Ordered by \code{constraintsVectMatrix}).}
  \item{@logLik}{-log-likeihood for models (ordered by \code{constraintsVectMatrix}).}
  \item{@aicc}{AICc for models (ordered by \code{constraintsVectMatrix}).}
  \item{@aic}{AIC for models (ordered by \code{constraintsVectMatrix}).}
  \item{@constraintsVectMatrix}{\code{constraintsVectMatrix} input to the function.}
  \item{@offsetVectMatrix}{\code{offsetVectMatrix} input to the function.}
  \item{@MLEs}{Maximum likelihood estimates for s parameters (ordered by \code{constraintsVectMatrix}).}
  \item{@SEs}{Standard errors for s parameters (ordered by \code{constraintsVectMatrix}).}
  \item{@MLEilv}{Maximum likelihood estimates for effect of ILs on asocial learning (ordered by
  \code{constraintsVectMatrix}).}
  \item{@SEilv}{Standard errors for effect of ILs on asocial learning  (ordered by \code{constraintsVectMatrix}).}
  \item{@MLEint}{Maximum likelihood estimates for effect of ILs on social learning (ordered by
  \code{constraintsVectMatrix}).}
  \item{@SEint}{Standard errors for effect of ILs on social learning  (ordered by \code{constraintsVectMatrix}).}
  \item{@MLEhaz}{Maximum likelihood estimates for baseline rate/hazard function (ordered by
  \code{constraintsVectMatrix}).}
  \item{@SEhaz}{Standard errors for baseline rate/hazard function (ordered by \code{constraintsVectMatrix}).}
  \item{@typeVect}{Type of models (ordered by \code{constraintsVectMatrix}).}
  \item{@baselineVect}{The baseline function used for each model. (ordered by \code{constraintsVectMatrix}).}
  \item{@deltaAICc}{Difference in AICc or AIC from the best model. (ordered by \code{constraintsVectMatrix}).}
  \item{@RelSupport}{Relative support for the model compared to the best model, calculated as exp(-0.5*deltaAICc).
  (ordered by \code{constraintsVectMatrix}).}
  \item{@RelSupport}{Akaike weight for the model. (ordered by \code{constraintsVectMatrix}).}
  \item{@printTable}{data.frame to be output by the print method for \code{tadaAICtable} (see above).}
}
}

\seealso{
\code{\link{networksSupport}}, \code{\link{typeByNetworksSupport}}, \code{\link{modelAverageEstimates}},
\code{\link{variableSupport}}, \code{\link{unconditionalStdErr}},\code{\link{baselineSupport}},
\code{\link{combineTadaAICtables}}. For OADA models use \code{\link{oadaAICtable}}.
}
