% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oadaAICtable.R
\name{typeByNetworksSupport}
\alias{typeByNetworksSupport}
\title{Get the support for each network x type combination from an oadaAICtable or tadaAICtable}
\usage{
typeByNetworksSupport(nbdaAICtable)
}
\arguments{
\item{nbdaAICtable}{an object of class \code{\link{oadaAICtable}} or \code{\link{tadaAICtable}}.}
}
\value{
array giving the support (total Akaike Weight) and number of models fitted for each network combination.
}
\description{
Calculates the support for network x type of model combination fitted in a set of models fitted using
\code{\link{oadaAICtable}} or \code{\link{tadaAICtable}}.
}
\details{
Network combinations are represented by the set of constraints on the s parameters used to generate the model, using the
\code{constraintsVectMatrix} argument. e.g. 1:1:1 means all three networks in the data were constrained to have the same
effect, 1:2:0 means the first two networks have different effects, whereas the third is constrained to have no effect, and
0:0:0 represents asocial models. For type, models are classified as: \describe{
  \item{noILVs;}{}
  \item{additive: ILV effects on asocial learning only;}{}
  \item{multiplicative: all ILVs have same effect on asocial and social learning;}{}
  \item{unconstrained: differing effects on asocial and social learning for at least one ILV;}{}
  \item{socialEffectsOnly: ILV effects only on social learning;}{}
  \item{asocial: no social transmission (all s=0).}{}
  }
}
\seealso{
\code{\link{oadaAICtable}}, \code{\link{tadaAICtable}}, \code{\link{typeSupport}},
\code{\link{typeByNetworksSupport}}, \code{\link{modelAverageEstimates}}, \code{\link{variableSupport}},
\code{\link{unconditionalStdErr}}.
}
