% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oadaAICtable.R
\name{typeSupport}
\alias{typeSupport}
\title{Get the support for each type of model from an oadaAICtable or tadaAICtable}
\usage{
typeSupport(nbdaAICtable)
}
\arguments{
\item{nbdaAICtable}{an object of class \code{\link{oadaAICtable}} or \code{\link{tadaAICtable}}.}
}
\value{
dataframe giving the support (total Akaike Weight) and number of models of each type.
}
\description{
Calculates the support for each type of model fitted in a set of models fitted using \code{\link{oadaAICtable}} or
\code{\link{tadaAICtable}}.
}
\details{
Models are classified as: \describe{
  \item{noILVs;}{}
  \item{additive: ILV effects on asocial learning only;}{}
  \item{multiplicative: all ILVs have same effect on asocial and social learning;}{}
  \item{unconstrained: differing effects on asocial and social learning for at least one ILV;}{}
  \item{socialEffectsOnly: ILV effects only on social learning;}{}
  \item{asocial: no social transmission (all s=0).}{}
  }
}
\seealso{
\code{\link{oadaAICtable}}, \code{\link{tadaAICtable}}, \code{\link{networksSupport}},
\code{\link{typeByNetworksSupport}}, \code{\link{modelAverageEstimates}}, \code{\link{variableSupport}},
\code{\link{unconditionalStdErr}}.
}
