% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oadaAICtable.R
\name{unconditionalStdErr}
\alias{unconditionalStdErr}
\title{Get the unconditional standard error for each variable from an oadaAICtable or tadaAICtable}
\usage{
unconditionalStdErr(nbdaAICtable, typeFilter = NULL, netFilter = NULL,
  baselineFilter = NULL, includeAsocial = TRUE, includeNoILVs = TRUE,
  nanReplace = FALSE)
}
\arguments{
\item{nbdaAICtable}{an object of class \code{\link{oadaAICtable}} or \code{\link{tadaAICtable}}.}

\item{typeFilter}{an optional string allowing the user to get the support for variables within a specified type of models.
e.g. \code{typeFilter="additive"} gets the variable support in the subset of additive models. See \code{\link{typeSupport}}
for an explanation of the different model types.}

\item{netFilter}{an optional string allowing the user to get the support for variables from models with a specific
combination of network effects, e.g. \code{netFilter = "0:1:0"} gets the variable support in the subset of models containing
only network 2. See \code{\link{networksSupport}} for an explanation of network combinations.}

\item{baselineFilter}{an optional string allowing the user to get the support for variables within the subset of models
with a specific baseline function, e.g. \code{typeFilter="gamma"} gets the variable support in the subset of models with a
gamma baseline rate function.
e.g. \code{typeFilter="additive"} gets the variable support in the subset of additive models.}

\item{includeAsocial}{logical indicating whether asocial models should be included.}

\item{includeNoILVs}{logical indicating whether models with no ILVs should be included.}

\item{nanReplace}{logical indicating whether standard errors recorded as NaNs should be replaced with the model averaged mean
SE across all models for which the SE for that variable could be calculated.}
}
\value{
numeric vector giving the model averaged estimate for each variable.
}
\description{
Calculates the unconditional standard error for each variable fitted in a set of models fitted using \code{\link{oadaAICtable}}
or code{\link{tadaAICtable}}.
}
\details{
A standard error (SE) from an individual model is conditional on that model. The unconditional standard error (USE) is not
really  "unconditional" but is rather conditional on the whole set of models fitted rather than an individual model. It takes
into account the SEs within each model and the variation in maximum likelihood estimates among models. SEs cannot be always be
calulated in an NBDA, meaning we are left with a situation in which USEs cannot be calculated if 1 or more models have SE=NaN-
even if there are only a few models of low weight for which this is the case. Here we offer a pragmatic
solution to such cases- by replacing the SEs in such models with a model-averaged SE across all models in which the SE can be
calculated, we are able to estimate a USE that is likely to be approximately correct. However, we recommend that this solution
only be used if SEs are absent for only a few models of low weight. In other cases we recommend that the user use confidence intervals
from \code{\link{profLikCI}} conditional on the best model as a measure of uncertainty in parameter estimates.
For s parameters \code{\link{multiModelLowerLimits}} and \code{\link{multiModelPropST}} then provide a means to assess the
robustness of the findings to model selection uncertainty.
}
\seealso{
\code{\link{oadaAICtable}}, \code{\link{tadaAICtable}}, \code{\link{typeSupport}},
\code{\link{typeByNetworksSupport}}, \code{\link{networksSupport}}, \code{\link{modelAverageEstimates}},
\code{\link{unconditionalStdErr}}.
}
