% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oadaAICtable.R
\name{variableSupport}
\alias{variableSupport}
\title{Get the support for each variable from an oadaAICtable or tadaAICtable}
\usage{
variableSupport(nbdaAICtable, typeFilter = NULL, baselineFilter = NULL,
  includeAsocial = TRUE)
}
\arguments{
\item{nbdaAICtable}{an object of class \code{\link{oadaAICtable}} or \code{\link{tadaAICtable}}.}

\item{typeFilter}{an optional string allowing the user to get the support for variables within a specified type of models.
e.g. \code{typeFilter="additive"} gets the variable support in the subset of additive models. See \code{\link{typeSupport}}
for an explanation of the different model types.}

\item{baselineFilter}{an optional string allowing the user to get the support for variables within the subset of models
with a specific baseline function, e.g. \code{typeFilter="gamma"} gets the variable support in the subset of models with a
gamma baseline rate function.
e.g. \code{typeFilter="additive"} gets the variable support in the subset of additive models.}

\item{includeAsocial}{logical indicating whether asocial models should be included.}
}
\value{
matrix giving the support (total Akaike Weight) for each variable.
}
\description{
Calculates the support for each variable fitted in a set of models fitted using \code{\link{oadaAICtable}} or
code{\link{tadaAICtable}}.
}
\details{
The support for each variable is the total Akaike weight for the models in which that variable is present, i.e. not
constrained to =0.
}
\seealso{
\code{\link{oadaAICtable}}, \code{\link{tadaAICtable}}, \code{\link{typeSupport}},
\code{\link{typeByNetworksSupport}}, \code{\link{networksSupport}}, \code{\link{modelAverageEstimates}},
\code{\link{unconditionalStdErr}}.
}
